feat: colored server status (#7)
### Description
This PR adds color to the server status when printing the services
statuses.

### Changes
- Sort `cargo.toml` dependencies
- Create ServerStatus enum to not deal with raw strings
- Implement `From` reqwest response result into ServerStatus
- Add method to return colored version of ServerStatus for displaying

### Notes
There are quite a few `*Status` structs on the `status.rs` file and it
takes a while to know which one is which and used where (makes even
harder since when assigning to variables, some places they are just
called `status`, which is kinda "hard" to know which of the status we
are talking about). Might be worth taking a look into some renames in
there.