refactor(cli): replace .unwrap where possible (#10)
Proper handle error where possible and where is not really an error that
we can/want to handle, replace it with expect so that we can easier
track down the failure.

From the [Rust
book](https://doc.rust-lang.org/book/ch09-02-recoverable-errors-with-result.html):

> (...) the expect method lets us also choose the panic! error message.
Using expect instead of unwrap and providing good error messages can
convey your intent and make tracking down the source of a panic easier.

> In production-quality code, most Rustaceans choose expect rather than
unwrap and give more context about why the operation is expected to
always succeed. That way, if your assumptions are ever proven wrong, you
have more information to use in debugging.