Fix overlapping route rewrites (#27)
Case is, target service on the remote side is a tunnel.
If the path gets rewritten once remotely, it can throw off finding the
right service in the local server.

Solution: an extra header linkup-destination set on every hop. If the
destination has been calculated once, we reuse it on later hops.