refactor: handle new lines to avoid interfering with existing variables
With this new change, when writing the new variables, Linkup will start by adding a new line to avoid writing the header on the same line as the latest variable.
On cleanup, it will also remove the wrapping new lines around the Linkup block (if any).

So, for example:
An existing config:
```
```

After start
```
##### Linkup environment - DO NOT EDIT #####
...
...
...
##### Linkup environment - DO NOT EDIT #####
```

After stop
```
```

Or an existing config:
```
a
b
c
```

After start
```
a
b
c
##### Linkup environment - DO NOT EDIT #####
...
...
...
##### Linkup environment - DO NOT EDIT #####

```

After stop
```
a
b
c
```

Or even if modified between start and stop, where existing is:
```
a
b
c
```

After start
```
a
b
c
##### Linkup environment - DO NOT EDIT #####
...
...
...
##### Linkup environment - DO NOT EDIT #####

```

Then modified to
```
a
b
c
##### Linkup environment - DO NOT EDIT #####
...
...
...
##### Linkup environment - DO NOT EDIT #####

d
e
```

After stop
```
a
b
c
d
e
```