refactor: extract env file handling (#32)
### Description
Move env file handling to own module so that we can better test its
behaviour and have it contained.

### Note
We need the service name on the env file handling just so that we can
use that while building the `CliError`. It feels that maybe we can
return another type of error by the env handlers and leave the
translation into a `CliError` as a responsibility of the caller. But
maybe we can do that on another PR.