refactor: shuffle code and responsibility of LocalState (#49)
### Description

This changes make so that we stop using methods from the `start` command
around the code.
To achieve this, the responsibility of loading and saving the state is
now part of the implementation
of the `LocalState` itself.

### Other notable changes
- Instead of passing around a Tuple for the local and remote
`StorableSession`, now there is a `ServerConfig` which
holds a `local` and `remote` fields. Where this struct is located could
be moved, but didn't gave that much thought.