refactor: remove dependency to async-trait
### Description
With the release of Rust [1.75](https://blog.rust-lang.org/2023/12/28/Rust-1.75.0.html#async-fn-and-return-position-impl-trait-in-traits), the first round of stabilization of async in traits is out! This does come with its own [set of caveats](https://blog.rust-lang.org/2023/12/21/async-fn-rpit-in-traits.html#where-the-gaps-lie). But those seem to be mostly for public traits.
The trait that we use it is, theoretically, public. But we intend it to be used by the other modules within the workspace, so it seems like it might be fine to move from the async-trait.

It might be good to test on the worker scenario, since there might be something there that I'm missing, but it seems by the tests and checks that is all working.