feat: revamp status command
This PR improves the feedback of the status command to live update the services responses. One "breaking change" is the removal of the --all/-a option.
With the idea to give better visibility of the overall status, that option doesn't make that much sense. (we can of course add it back if we think is very important here).

Changelog:
- Removed --all/-a options. Now all services are shown by default.
- Added dependency to [`crossterm`](https://docs.rs/crossterm/latest/crossterm/) to manipulate the terminal output.
- Cleaned up the output (removed "Service information:" and "Session information:" headers).
- Services statuses are update as soon as we get them.
- Removed `component_kind: linkup`. Internal Linkup services are reported as remote or local accordingly.
- Added priority option for component status. This way we can keep the Linkup internal services reported on the beginning of the list.