feat: create health command (#116)
We have very often to instruct people to do some steps and verify some
local settings so that we can have an idea of what might be wrong with
their environment.
It would be good for Linkup to have a command that output the overall
health so that we can just get the dump from this command and have an
idea of what might be wrong.

### Example of output
```
System info:
  OS: Darwin (14.6.1)
Session info:
  Name: sandy-camel
Background sevices:
  - Caddy       RUNNING (76550)
  - dnsmasq     RUNNING (76565)
  - Cloudflared RUNNING (76530)
Environment variables:
  - LINKUP_CF_API_TOKEN          OK
  - LINKUP_CLOUDFLARE_ZONE_ID    OK
  - LINKUP_CLOUDFLARE_ACCOUNT_ID OK
Linkup:
  Version: 1.2.1
  Config location: $HOME/.linkup
  Config contents:
    - dnsmasq-log
    - localserver-stdout
    - dnsmasq-pid
    - localserver-pid
    - cloudflared-pid
    - localserver-stderr
    - state
    - cloudflared-stdout
    - Caddyfile
    - localdns-install
    - caddy-log
    - dnsmasq-conf
    - cloudflared-stderr
    - caddy-pid
Local DNS resolvers:
    - <resolver_1>
    - <resolver_2>
```