feat: add command to start Linkup local server
This new command/feature allowed us to be able to remove daemonize and fix errors regarding process forking.
With this, now the cli has a command that simply starts the server, and with that, the CLI can manage
the server process the same way it manages caddy, cloudflared and dnsmasq.

For this to work, the CLI needs to support Tokio so it can run the server. This caused a cascade of necessary
changes to support async.