refactor: rework background services (#121)
### Description
These changes introduce a preview of how the start process is going when
using the `linkup start` command.

### New start feedback


https://github.com/user-attachments/assets/5e0ea244-aa34-46af-8f9d-d5d98a42e96b

### Changelog
- Linkup CLI now spins up local-server as a subcommand. This allows us
to remove the dependency on `Daemonize`, but we need to depend on Tokio
now. But now the local-server is ran like all the other background
services with `std::process:Command`.
- `start` command now has a live update of the steps of the process that
are going on.
- Add `tunnel url` to the `health` command.

### TODO
- [x] Block double start.

### Thoughts
- Maybe we can have a `next` branch and create beta/rc releases from it
before generating the widely available one?

---------

Co-authored-by: Oliver Stenbom <oliver@stenbom.eu>