fix: look for name including file name
There should be only one file inside the archive, we do the name check just for sanity.
The reason why we can't look for exactly the same name anymore is because the Caddy executables
are saved also with the architecture and OS on the name of the file inside the archive.