fix: allow HTTPS server to run without certificates (#207)
Currently, if the certificates folder does not exist, the HTTPS server
fails to start. With the current setup, if any of the two servers (HTTP
or HTTPS) exits, it stops the other one.

This is causing issue where, if the user does not have local-dns
installed, they won't have the certificates folder, and by consequence
the HTTPS server will not start and kill the HTTP server.

These changes allow the HTTPS server to start without any certificate.