refactor: change from global error enum to anyhow
This changes to use (anyhow)[https://docs.rs/anyhow/latest/anyhow/] instead of a "global" error enum.
This does not fully tackle adding context to places where errors happens, but more replace on places where it was straightforward to change one for the other.

As a follow-up, should go on places that we currently have `.unwrap()`, `.expect()` and some `?` to add `.context()/.with_context()` to them.