fix: possible orphans ignore linkup folder path if is in env variable (#225)
Some processes add the linkup bin folder somewhere on their command.
Sometimes as environment variable.

For example, `bundle exec ...` load the users `PATH` into a variable
`BUNDLER_ORIG_PATH`, which contains the path to the linkup bin folder.
These should not be considered possible orphan processes, but they are
at the moment.

Also now maxing the size of the output of the orphan commands to be 120
characters to avoid huge outputs when the list of arguments are too
long.