fix: Linux installation lifecycle (#250)
### Description
We are currently installing Linkup under `$HOME/.linkup/bin` for both
Linux and MacOS.
The issue with this is that on Linux, we need to use `setcap` to allow
the process to bind to process 80/443, which does not work if installed
on that location.
On Linux, file capabilities (setcap) only work reliably on files located
on a filesystem that supports extended attributes and is mounted with
them enabled.

This changes so that we now install on Linux under `/usr/local/bin`.
This requires sudo access on installation, uninstallation and update.

This should unblock us to merge #216 

Closes SHIP-2224